Broke StoreProduct into 3 components to use them separately and provided minor additions.
