+++ client/src/App.js
@@ -1,10 +1,11 @@
 import SignIn from './components/SignIn';
+import StoreProducts from './components/StoreProducts';
 import './App.css';
 
 function App() {
   return (
     <div>
-      <SignIn />
+      <StoreProducts />
     </div>
   );
 }
